+++ nexus/social_media/services.py
@@ -1,14 +1,169 @@
 # -*- coding: utf-8 -*-
 # Third Party Stuff
 import facebook
+import requests
 import tweepy
 from django.conf import settings
 from django.utils import timezone
+from rest_framework import status
 
 # nexus Stuff
-from nexus.base import exceptions
+from nexus.base import exceptions as exc
 from nexus.social_media import tasks
 from nexus.social_media.models import Post
+
+
+def upload_image_to_linkedin(author, headers, post, linkedin_auth):
+    """Upload image to linkedin for given post.
+
+    :param author: Owner of the post to be used for response POST data.
+
+    :param headers: HTTP headers required for successful HTTP request.
+
+    :param post: Post model instance.
+
+    :param linkedin_auth: LINKEDIN_AUTH object taken from settings.
+
+    :returns: The value of asset provided all responses are successful.
+
+    """
+    post_data_assets = {
+        'registerUploadRequest': {
+            'recipes': [
+                'urn:li:digitalmediaRecipe:feedshare-image'
+            ],
+            'owner': author,
+            'serviceRelationships': [
+                {
+                    'relationshipType': 'OWNER',
+                    'identifier': 'urn:li:userGeneratedContent'
+                }
+            ]
+        }
+    }
+
+    # LinkedIn assets API endpoint
+    api_url_assets = f'{settings.LINKEDIN_API_URL_BASE}assets?action=registerUpload'
+
+    response_assets = requests.post(api_url_assets, json=post_data_assets,
+                                    headers=headers)
+    if response_assets.status_code == status.HTTP_200_OK:
+        json_response_assets = response_assets.json()
+
+        value = json_response_assets.get('value')
+        upload_mechanism = value.get('uploadMechanism')
+        upload_http_request = upload_mechanism.get('com.linkedin.digitalmedia.uploading.MediaUploadHttpRequest')
+        upload_url = upload_http_request.get('uploadUrl')
+
+        asset = value.get('asset')
+
+        response_upload_image = requests.post(upload_url,
+                                              data=post.image.file.read(),
+                                              headers={'Authorization': f"Bearer {linkedin_auth['access_token']}"})
+        if response_upload_image.status_code == status.HTTP_201_CREATED:
+            return asset
+        else:
+            check_and_raise_error_from_linkedin_response(response_upload_image)
+    else:
+        check_and_raise_error_from_linkedin_response(response_assets)
+
+
+def publish_on_linkedin(post_id):
+    """Function to post on linkedin
+
+    :param post_id: UUID of the post instance to be posted.
+
+    :returns: A valid HTTP response
+
+    """
+    linkedin_auth = settings.LINKEDIN_AUTH
+    author = f"urn:li:organization:{linkedin_auth['organization_id']}"
+
+    headers = {'X-Restli-Protocol-Version': '2.0.0',
+               'Content-Type': 'application/json',
+               'Authorization': f"Bearer {linkedin_auth['access_token']}"}
+
+    # LinkedIn UGC(User Generated Content) API endpoint
+    api_url_ugc = f'{settings.LINKEDIN_API_URL_BASE}ugcPosts'
+
+    post = Post.objects.get(pk=post_id)
+
+    post_data = {
+        'author': author,
+        'lifecycleState': 'PUBLISHED',
+        'specificContent': {
+            'com.linkedin.ugc.ShareContent': {
+                'shareCommentary': {
+                    'text': post.text
+                },
+                'shareMediaCategory': 'NONE'
+            },
+        },
+        'visibility': {
+            'com.linkedin.ugc.MemberNetworkVisibility': 'PUBLIC'
+        },
+    }
+
+    if post.image:
+
+        # asset is required to include the image in our post after successful uploading of image.
+        asset = upload_image_to_linkedin(author, headers, post, linkedin_auth)
+
+        specific_content = post_data.get('specificContent')
+        share_content = specific_content.get('com.linkedin.ugc.ShareContent')
+
+        image_media = [{
+            'status': 'READY',
+            'description': {
+                'text': ''
+            },
+            'media': asset,
+            'title': {
+                'text': ''
+            }
+        }]
+        share_content.update(shareMediaCategory='IMAGE',
+                             media=image_media)
+        image_response = requests.post(api_url_ugc, headers=headers,
+                                       json=post_data)
+        return image_response
+    elif post.text:
+        response = requests.post(api_url_ugc, headers=headers, json=post_data)
+        return response
+
+
+def check_and_raise_error_from_linkedin_response(response):
+    """Function to perform appropriate action based on the linkedin response given.
+
+    :param response: A valid HTTP response to perform suitable action.
+
+    :raises WrongArguments: Exception raised when either a 400 or 404 response is provided.
+
+    :raises NotAuthenticated: Exception raised when a 401 response is provided.
+
+    :raises PermissionDenied: Exception raised when a 403 response is provided.
+
+    :raises NotSupported: Exception raised when a 405 response is provided.
+
+    :raises RequestValidationError: Exception raised when an unexpected response is provided.
+
+    """
+    status_code = response.status_code
+    if (status_code != status.HTTP_201_CREATED) or (status_code != status.HTTP_200_OK):
+        json_response = response.json()
+        error = f"LinkedIn error: {json_response['message']}"
+        if status_code == status.HTTP_400_BAD_REQUEST:
+            raise exc.WrongArguments(error)
+        elif status_code == status.HTTP_401_UNAUTHORIZED:
+            raise exc.NotAuthenticated(error)
+        elif status_code == status.HTTP_403_FORBIDDEN:
+            raise exc.PermissionDenied(error)
+        elif status_code == status.HTTP_404_NOT_FOUND:
+            raise exc.WrongArguments(error)
+        elif status_code == status.HTTP_405_METHOD_NOT_ALLOWED:
+            raise exc.NotSupported(error)
+        else:
+            raise exc.RequestValidationError(error)
 
 
 def get_twitter_api_object(TWITTER_OAUTH):
@@ -27,8 +182,8 @@ def get_twitter_api_object(TWITTER_OAUTH):
         auth.set_access_token(TWITTER_OAUTH['access_key'], TWITTER_OAUTH['access_secret'])
         twitter_api = tweepy.API(auth)
         return twitter_api
-    except tweepy.error.TweepError as exc:
-        raise exceptions.WrongArguments(str(exc))
+    except tweepy.error.TweepError as exc_info:
+        raise exc.WrongArguments(str(exc_info))
 
 
 def publish_on_twitter(post_id):
@@ -48,9 +203,8 @@ def publish_on_twitter(post_id):
             twitter_api.update_with_media(filename=filename, status=post.text, file=post.image)
         elif post.text:
             twitter_api.update_status(status=post.text)
-
-    except tweepy.error.TweepError as exc:
-        raise exceptions.BadRequest(str(exc))
+    except tweepy.error.TweepError as exc_info:
+        raise exc.BadRequest(str(exc_info))
 
 
 def get_fb_page_graph():
@@ -60,9 +214,9 @@ def get_fb_page_graph():
     try:
         page_list = list(filter(lambda page: page['id'] == settings.FB_PAGE_ID, pages))
     except KeyError:
-        raise exceptions.WrongArguments('No ID associated with FB_USER_ACCESS_TOKEN.')
+        raise exc.WrongArguments('No ID associated with FB_USER_ACCESS_TOKEN.')
     if not page_list:
-        raise exceptions.WrongArguments('No matching ID in account data. Incorrect FB_PAGE_ID')
+        raise exc.WrongArguments('No matching ID in account data. Incorrect FB_PAGE_ID')
     page_access_token = page_list[0]['access_token']
     page_graph = facebook.GraphAPI(page_access_token)
     return page_graph
@@ -73,7 +227,8 @@ def publish_on_facebook(post_id):
     page_graph = get_fb_page_graph()
     if post.image:
         if post.text:
-            page_graph.put_photo(image=post.image.file.open('rb'), message=post.text)
+            page_graph.put_photo(image=post.image.file.open('rb'),
+                                 message=post.text)
         else:
             page_graph.put_photo(image=post.image.file.open('rb'))
     elif post.text:
@@ -99,3 +254,5 @@ def publish_on_social_media():
             tasks.publish_on_facebook_task.delay(post['id'])
         elif post['posted_at'] == 'twitter':
             tasks.publish_on_twitter_task.delay(post['id'])
+        elif post['posted_at'] == 'linkedin':
+            tasks.publish_on_linkedin_task.delay(post['id'])