+++ docs/api/changelog.md
@@ -4,6 +4,10 @@
     .container hr{margin-top: 5px; }
 </style>
 
+## Aug 9, 2019
+
+- Add feature to publish social media post on linkedin ([@gutsytechster])
+
 ## Aug 8, 2019
 
 - Add feature to publish social media post on twitter ([@GeekyShacklebolt])